Переместил все папки с классами в одну общую
