+++ Tiere/Pets/Dogs/Dogs.java
@@ -1,6 +1,6 @@
-package Pets.Dogs;
+package Tiere.Pets.Dogs;
 
-import Pets.Pets;
+import Tiere.Pets.Pets;
 
 public class Dogs extends Pets {
     public Dogs(String name, String dateOfBirth, String[] commands) {
